+++ ignite/templates/typed/dry/dry.go
@@ -9,16 +9,16 @@ import (
 	"github.com/ignite/cli/ignite/templates/typed"
 )
 
-//go:embed stargate/component/* stargate/component/**/*
-var fsStargateComponent embed.FS
+//go:embed files/component/* files/component/**/*
+var fsComponent embed.FS
 
 // NewStargate returns the generator to scaffold a basic type in a Stargate module.
 func NewStargate(opts *typed.Options) (*genny.Generator, error) {
 	var (
 		g        = genny.New()
 		template = xgenny.NewEmbedWalker(
-			fsStargateComponent,
-			"stargate/component/",
+			fsComponent,
+			"files/component/",
 			opts.AppPath,
 		)
 	)
