+++ ignite/templates/typed/dry/files/component/proto/{{appName}}/{{moduleName}}/{{typeName}}.proto.plush
