+++ ignite/templates/typed/list/files/component/proto/{{appName}}/{{moduleName}}/{{typeName}}.proto.plush
