+++ ignite/templates/typed/list/files/component/x/{{moduleName}}/client/cli/query_{{typeName}}.go.plush
