+++ ignite/templates/typed/list/files/component/x/{{moduleName}}/keeper/grpc_query_{{typeName}}.go.plush
