+++ ignite/templates/typed/list/files/component/x/{{moduleName}}/keeper/{{typeName}}.go.plush
