+++ ignite/templates/typed/list/files/messages/x/{{moduleName}}/client/cli/tx_{{typeName}}.go.plush
