+++ ignite/templates/typed/list/files/messages/x/{{moduleName}}/types/messages_{{typeName}}_test.go.plush
