+++ ignite/templates/typed/map/files/component/proto/{{appName}}/{{moduleName}}/{{typeName}}.proto.plush
