+++ ignite/templates/typed/map/files/component/x/{{moduleName}}/client/cli/query_{{typeName}}.go.plush
