+++ ignite/templates/typed/map/files/component/x/{{moduleName}}/keeper/grpc_query_{{typeName}}.go.plush
