+++ ignite/templates/typed/map/files/component/x/{{moduleName}}/keeper/{{typeName}}.go.plush
