+++ ignite/templates/typed/map/files/component/x/{{moduleName}}/types/key_{{typeName}}.go.plush
