+++ ignite/templates/typed/map/files/messages/x/{{moduleName}}/client/cli/tx_{{typeName}}.go.plush
