+++ ignite/templates/typed/map/files/messages/x/{{moduleName}}/types/messages_{{typeName}}_test.go.plush
