+++ ignite/templates/typed/map/files/tests/component/x/{{moduleName}}/keeper/grpc_query_{{typeName}}_test.go.plush
