+++ ignite/templates/typed/singleton/files/component/proto/{{appName}}/{{moduleName}}/{{typeName}}.proto.plush
