+++ ignite/templates/typed/singleton/files/component/x/{{moduleName}}/client/cli/query_{{typeName}}.go.plush
