+++ ignite/templates/typed/singleton/files/component/x/{{moduleName}}/keeper/grpc_query_{{typeName}}_test.go.plush
