+++ ignite/templates/typed/singleton/files/component/x/{{moduleName}}/keeper/{{typeName}}.go.plush
