+++ ignite/templates/typed/singleton/files/component/x/{{moduleName}}/keeper/{{typeName}}_test.go.plush
