+++ ignite/templates/typed/singleton/files/messages/x/{{moduleName}}/types/messages_{{typeName}}_test.go.plush
