+++ ignite/templates/typed/singleton/singleton.go
@@ -1,8 +1,9 @@
-package maptype
+package singleton
 
 import (
 	"embed"
 	"fmt"
+	"math/rand"
 	"path/filepath"
 	"strings"
 
@@ -11,69 +12,44 @@ import (
 	"github.com/ignite/cli/ignite/pkg/gomodulepath"
 	"github.com/ignite/cli/ignite/pkg/placeholder"
 	"github.com/ignite/cli/ignite/pkg/xgenny"
-	"github.com/ignite/cli/ignite/templates/field/datatype"
 	"github.com/ignite/cli/ignite/templates/module"
 	"github.com/ignite/cli/ignite/templates/typed"
 )
 
 var (
-	//go:embed stargate/component/* stargate/component/**/*
-	fsStargateComponent embed.FS
+	//go:embed files/messages/* files/messages/**/*
+	fsMessages embed.FS
 
-	//go:embed stargate/messages/* stargate/messages/**/*
-	fsStargateMessages embed.FS
+	//go:embed files/component/* files/component/**/*
+	fsComponent embed.FS
 
-	//go:embed stargate/tests/component/* stargate/tests/component/**/*
-	fsStargateTestsComponent embed.FS
-
-	//go:embed stargate/tests/messages/* stargate/tests/messages/**/*
-	fsStargateTestsMessages embed.FS
-
-	//go:embed stargate/simapp/* stargate/simapp/**/*
-	fsStargateSimapp embed.FS
+	//go:embed files/simapp/* files/simapp/**/*
+	fsimapp embed.FS
 )
 
-// NewStargate returns the generator to scaffold a new map type in a Stargate module
+// NewStargate returns the generator to scaffold a new indexed type in a Stargate module
 func NewStargate(replacer placeholder.Replacer, opts *typed.Options) (*genny.Generator, error) {
-	// Tests are not generated for map with a custom index that contains only booleans
-	// because we can't generate reliable tests for this type
-	var generateTest bool
-	for _, index := range opts.Indexes {
-		if index.DatatypeName != datatype.Bool {
-			generateTest = true
-		}
-	}
-
 	var (
 		g = genny.New()
 
 		messagesTemplate = xgenny.NewEmbedWalker(
-			fsStargateMessages,
-			"stargate/messages/",
-			opts.AppPath,
-		)
-		testsMessagesTemplate = xgenny.NewEmbedWalker(
-			fsStargateTestsMessages,
-			"stargate/tests/messages/",
+			fsMessages,
+			"files/messages/",
 			opts.AppPath,
 		)
 		componentTemplate = xgenny.NewEmbedWalker(
-			fsStargateComponent,
-			"stargate/component/",
-			opts.AppPath,
-		)
-		testsComponentTemplate = xgenny.NewEmbedWalker(
-			fsStargateTestsComponent,
-			"stargate/tests/component/",
+			fsComponent,
+			"files/component/",
 			opts.AppPath,
 		)
 		simappTemplate = xgenny.NewEmbedWalker(
-			fsStargateSimapp,
-			"stargate/simapp/",
+			fsimapp,
+			"files/simapp/",
 			opts.AppPath,
 		)
 	)
 
+	g.RunFn(typesKeyModify(opts))
 	g.RunFn(protoRPCModify(replacer, opts))
 	g.RunFn(clientCliQueryModify(replacer, opts))
 	g.RunFn(genesisProtoModify(replacer, opts))
@@ -98,19 +74,26 @@ func NewStargate(replacer placeholder.Replacer, opts *typed.Options) (*genny.Gen
 		if err := typed.Box(messagesTemplate, opts, g); err != nil {
 			return nil, err
 		}
-		if generateTest {
-			if err := typed.Box(testsMessagesTemplate, opts, g); err != nil {
-				return nil, err
-			}
-		}
 	}
 
-	if generateTest {
-		if err := typed.Box(testsComponentTemplate, opts, g); err != nil {
-			return nil, err
-		}
-	}
 	return g, typed.Box(componentTemplate, opts, g)
+}
+
+func typesKeyModify(opts *typed.Options) genny.RunFn {
+	return func(r *genny.Runner) error {
+		path := filepath.Join(opts.AppPath, "x", opts.ModuleName, "types/keys.go")
+		f, err := r.Disk.Find(path)
+		if err != nil {
+			return err
+		}
+		content := f.String() + fmt.Sprintf(`
+const (
+	%[1]vKey= "%[1]v/value/"
+)
+`, opts.TypeName.UpperCamel)
+		newFile := genny.NewFileS(path, content)
+		return r.File(newFile)
+	}
 }
 
 func protoRPCModify(replacer placeholder.Replacer, opts *typed.Options) genny.RunFn {
@@ -136,79 +119,31 @@ func protoRPCModify(replacer placeholder.Replacer, opts *typed.Options) genny.Ru
 		replacementGogoImport := typed.EnsureGogoProtoImported(path, typed.Placeholder)
 		content = replacer.Replace(content, typed.Placeholder, replacementGogoImport)
 
-		var protoIndexes []string
-		for _, index := range opts.Indexes {
-			protoIndexes = append(protoIndexes, fmt.Sprintf("{%s}", index.ProtoFieldName()))
-		}
-		indexPath := strings.Join(protoIndexes, "/")
-
 		// Add the service
 		templateService := `// Queries a %[2]v by index.
 	rpc %[2]v(QueryGet%[2]vRequest) returns (QueryGet%[2]vResponse) {
-		option (google.api.http).get = "/%[3]v/%[4]v/%[5]v/%[6]v";
-	}
-
-	// Queries a list of %[2]v items.
-	rpc %[2]vAll(QueryAll%[2]vRequest) returns (QueryAll%[2]vResponse) {
 		option (google.api.http).get = "/%[3]v/%[4]v/%[5]v";
 	}
-
 %[1]v`
 		appModulePath := gomodulepath.ExtractAppPath(opts.ModulePath)
-		replacementService := fmt.Sprintf(templateService,
-			typed.Placeholder2,
+		replacementService := fmt.Sprintf(templateService, typed.Placeholder2,
 			opts.TypeName.UpperCamel,
 			appModulePath,
 			opts.ModuleName,
 			opts.TypeName.Snake,
-			indexPath,
 		)
 		content = replacer.Replace(content, typed.Placeholder2, replacementService)
 
 		// Add the service messages
-		var queryIndexFields string
-		for i, index := range opts.Indexes {
-			queryIndexFields += fmt.Sprintf("  %s;\n", index.ProtoType(i+1))
-		}
-
-		// Ensure custom types are imported
-		protoImports := opts.Fields.ProtoImports()
-		for _, f := range opts.Fields.Custom() {
-			protoImports = append(protoImports,
-				fmt.Sprintf("%[1]v/%[2]v/%[3]v.proto", opts.AppName, opts.ModuleName, f),
-			)
-		}
-		for _, f := range protoImports {
-			importModule := fmt.Sprintf(`
-import "%[1]v";`, f)
-			content = strings.ReplaceAll(content, importModule, "")
-			replacementImport := fmt.Sprintf("%[1]v%[2]v", typed.Placeholder, importModule)
-			content = replacer.Replace(content, typed.Placeholder, replacementImport)
-		}
-
-		templateMessage := `message QueryGet%[2]vRequest {
-	%[4]v
-}
+		templateMessage := `message QueryGet%[2]vRequest {}
 
 message QueryGet%[2]vResponse {
-	%[2]v %[3]v = 1 [(gogoproto.nullable) = false];
+	%[2]v %[2]v = 1 [(gogoproto.nullable) = false];
 }
-
-message QueryAll%[2]vRequest {
-	cosmos.base.query.v1beta1.PageRequest pagination = 1;
-}
-
-message QueryAll%[2]vResponse {
-	repeated %[2]v %[3]v = 1 [(gogoproto.nullable) = false];
-	cosmos.base.query.v1beta1.PageResponse pagination = 2;
-}
-
 %[1]v`
-		replacementMessage := fmt.Sprintf(templateMessage,
-			typed.Placeholder3,
+		replacementMessage := fmt.Sprintf(templateMessage, typed.Placeholder3,
 			opts.TypeName.UpperCamel,
 			opts.TypeName.LowerCamel,
-			queryIndexFields,
 		)
 		content = replacer.Replace(content, typed.Placeholder3, replacementMessage)
 
@@ -224,8 +159,7 @@ func clientCliQueryModify(replacer placeholder.Replacer, opts *typed.Options) ge
 		if err != nil {
 			return err
 		}
-		template := `cmd.AddCommand(CmdList%[2]v())
-	cmd.AddCommand(CmdShow%[2]v())
+		template := `cmd.AddCommand(CmdShow%[2]v())
 %[1]v`
 		replacement := fmt.Sprintf(template, typed.Placeholder,
 			opts.TypeName.UpperCamel,
@@ -255,17 +189,13 @@ func genesisProtoModify(replacer placeholder.Replacer, opts *typed.Options) genn
 		)
 		content := replacer.Replace(f.String(), typed.PlaceholderGenesisProtoImport, replacementProtoImport)
 
-		// Add gogo.proto
-		replacementGogoImport := typed.EnsureGogoProtoImported(path, typed.PlaceholderGenesisProtoImport)
-		content = replacer.Replace(content, typed.PlaceholderGenesisProtoImport, replacementGogoImport)
-
 		// Parse proto file to determine the field numbers
 		highestNumber, err := typed.GenesisStateHighestFieldNumber(path)
 		if err != nil {
 			return err
 		}
 
-		templateProtoState := `repeated %[2]v %[3]vList = %[4]v [(gogoproto.nullable) = false];
+		templateProtoState := `%[2]v %[3]v = %[4]v;
   %[1]v`
 		replacementProtoState := fmt.Sprintf(
 			templateProtoState,
@@ -291,10 +221,7 @@ func genesisTypesModify(replacer placeholder.Replacer, opts *typed.Options) genn
 
 		content := typed.PatchGenesisTypeImport(replacer, f.String())
 
-		templateTypesImport := `"fmt"`
-		content = replacer.ReplaceOnce(content, typed.PlaceholderGenesisTypesImport, templateTypesImport)
-
-		templateTypesDefault := `%[2]vList: []%[2]v{},
+		templateTypesDefault := `%[2]v: nil,
 %[1]v`
 		replacementTypesDefault := fmt.Sprintf(
 			templateTypesDefault,
@@ -302,69 +229,6 @@ func genesisTypesModify(replacer placeholder.Replacer, opts *typed.Options) genn
 			opts.TypeName.UpperCamel,
 		)
 		content = replacer.Replace(content, typed.PlaceholderGenesisTypesDefault, replacementTypesDefault)
-
-		// lines of code to call the key function with the indexes of the element
-		var indexArgs []string
-		for _, index := range opts.Indexes {
-			indexArgs = append(indexArgs, "elem."+index.Name.UpperCamel)
-		}
-		keyCall := fmt.Sprintf("%sKey(%s)", opts.TypeName.UpperCamel, strings.Join(indexArgs, ","))
-
-		templateTypesValidate := `// Check for duplicated index in %[2]v
-%[2]vIndexMap := make(map[string]struct{})
-
-for _, elem := range gs.%[3]vList {
-	index := %[4]v
-	if _, ok := %[2]vIndexMap[index]; ok {
-		return fmt.Errorf("duplicated index for %[2]v")
-	}
-	%[2]vIndexMap[index] = struct{}{}
-}
-%[1]v`
-		replacementTypesValidate := fmt.Sprintf(
-			templateTypesValidate,
-			typed.PlaceholderGenesisTypesValidate,
-			opts.TypeName.LowerCamel,
-			opts.TypeName.UpperCamel,
-			fmt.Sprintf("string(%s)", keyCall),
-		)
-		content = replacer.Replace(content, typed.PlaceholderGenesisTypesValidate, replacementTypesValidate)
-
-		newFile := genny.NewFileS(path, content)
-		return r.File(newFile)
-	}
-}
-
-func genesisModuleModify(replacer placeholder.Replacer, opts *typed.Options) genny.RunFn {
-	return func(r *genny.Runner) error {
-		path := filepath.Join(opts.AppPath, "x", opts.ModuleName, "genesis.go")
-		f, err := r.Disk.Find(path)
-		if err != nil {
-			return err
-		}
-
-		templateModuleInit := `// Set all the %[2]v
-for _, elem := range genState.%[3]vList {
-	k.Set%[3]v(ctx, elem)
-}
-%[1]v`
-		replacementModuleInit := fmt.Sprintf(
-			templateModuleInit,
-			typed.PlaceholderGenesisModuleInit,
-			opts.TypeName.LowerCamel,
-			opts.TypeName.UpperCamel,
-		)
-		content := replacer.Replace(f.String(), typed.PlaceholderGenesisModuleInit, replacementModuleInit)
-
-		templateModuleExport := `genesis.%[3]vList = k.GetAll%[3]v(ctx)
-%[1]v`
-		replacementModuleExport := fmt.Sprintf(
-			templateModuleExport,
-			typed.PlaceholderGenesisModuleExport,
-			opts.TypeName.LowerCamel,
-			opts.TypeName.UpperCamel,
-		)
-		content = replacer.Replace(content, typed.PlaceholderGenesisModuleExport, replacementModuleExport)
 
 		newFile := genny.NewFileS(path, content)
 		return r.File(newFile)
@@ -379,31 +243,24 @@ func genesisTestsModify(replacer placeholder.Replacer, opts *typed.Options) genn
 			return err
 		}
 
-		// Create a list of two different indexes to use as sample
-		sampleIndexes := make([]string, 2)
-		for i := 0; i < 2; i++ {
-			for _, index := range opts.Indexes {
-				sampleIndexes[i] += index.GenesisArgs(i)
-			}
+		// Create a fields
+		sampleFields := ""
+		for _, field := range opts.Fields {
+			sampleFields += field.GenesisArgs(rand.Intn(100) + 1)
 		}
 
-		templateState := `%[2]vList: []types.%[2]v{
-		{
-			%[3]v},
-		{
-			%[4]v},
-	},
-	%[1]v`
+		templateState := `%[2]v: &types.%[2]v{
+		%[3]v},
+		%[1]v`
 		replacementState := fmt.Sprintf(
 			templateState,
 			module.PlaceholderGenesisTestState,
 			opts.TypeName.UpperCamel,
-			sampleIndexes[0],
-			sampleIndexes[1],
+			sampleFields,
 		)
 		content := replacer.Replace(f.String(), module.PlaceholderGenesisTestState, replacementState)
 
-		templateAssert := `require.ElementsMatch(t, genesisState.%[2]vList, got.%[2]vList)
+		templateAssert := `require.Equal(t, genesisState.%[2]v, got.%[2]v)
 %[1]v`
 		replacementTests := fmt.Sprintf(
 			templateAssert,
@@ -425,51 +282,62 @@ func genesisTypesTestsModify(replacer placeholder.Replacer, opts *typed.Options)
 			return err
 		}
 
-		// Create a list of two different indexes to use as sample
-		sampleIndexes := make([]string, 2)
-		for i := 0; i < 2; i++ {
-			for _, index := range opts.Indexes {
-				sampleIndexes[i] += index.GenesisArgs(i)
-			}
+		// Create a fields
+		sampleFields := ""
+		for _, field := range opts.Fields {
+			sampleFields += field.GenesisArgs(rand.Intn(100) + 1)
 		}
 
-		templateValid := `%[2]vList: []types.%[2]v{
-	{
+		templateValid := `%[2]v: &types.%[2]v{
 		%[3]v},
-	{
-		%[4]v},
-},
 %[1]v`
 		replacementValid := fmt.Sprintf(
 			templateValid,
 			module.PlaceholderTypesGenesisValidField,
 			opts.TypeName.UpperCamel,
-			sampleIndexes[0],
-			sampleIndexes[1],
+			sampleFields,
 		)
 		content := replacer.Replace(f.String(), module.PlaceholderTypesGenesisValidField, replacementValid)
 
-		templateDuplicated := `{
-	desc:     "duplicated %[2]v",
-	genState: &types.GenesisState{
-		%[3]vList: []types.%[3]v{
-			{
-				%[4]v},
-			{
-				%[4]v},
-		},
-	},
-	valid:    false,
-},
+		newFile := genny.NewFileS(path, content)
+		return r.File(newFile)
+	}
+}
+
+func genesisModuleModify(replacer placeholder.Replacer, opts *typed.Options) genny.RunFn {
+	return func(r *genny.Runner) error {
+		path := filepath.Join(opts.AppPath, "x", opts.ModuleName, "genesis.go")
+		f, err := r.Disk.Find(path)
+		if err != nil {
+			return err
+		}
+
+		templateModuleInit := `// Set if defined
+if genState.%[3]v != nil {
+	k.Set%[3]v(ctx, *genState.%[3]v)
+}
 %[1]v`
-		replacementDuplicated := fmt.Sprintf(
-			templateDuplicated,
-			module.PlaceholderTypesGenesisTestcase,
+		replacementModuleInit := fmt.Sprintf(
+			templateModuleInit,
+			typed.PlaceholderGenesisModuleInit,
 			opts.TypeName.LowerCamel,
 			opts.TypeName.UpperCamel,
-			sampleIndexes[0],
 		)
-		content = replacer.Replace(content, module.PlaceholderTypesGenesisTestcase, replacementDuplicated)
+		content := replacer.Replace(f.String(), typed.PlaceholderGenesisModuleInit, replacementModuleInit)
+
+		templateModuleExport := `// Get all %[2]v
+%[2]v, found := k.Get%[3]v(ctx)
+if found {
+	genesis.%[3]v = &%[2]v
+}
+%[1]v`
+		replacementModuleExport := fmt.Sprintf(
+			templateModuleExport,
+			typed.PlaceholderGenesisModuleExport,
+			opts.TypeName.LowerCamel,
+			opts.TypeName.UpperCamel,
+		)
+		content = replacer.Replace(content, typed.PlaceholderGenesisModuleExport, replacementModuleExport)
 
 		newFile := genny.NewFileS(path, content)
 		return r.File(newFile)
@@ -506,20 +374,14 @@ func protoTxModify(replacer placeholder.Replacer, opts *typed.Options) genny.Run
 		content = replacer.Replace(content, typed.PlaceholderProtoTxRPC, replacementRPC)
 
 		// Messages
-		var indexes string
-		for i, index := range opts.Indexes {
-			indexes += fmt.Sprintf("  %s;\n", index.ProtoType(i+2))
-		}
-
 		var fields string
-		for i, f := range opts.Fields {
-			fields += fmt.Sprintf("  %s;\n", f.ProtoType(i+2+len(opts.Indexes)))
+		for i, field := range opts.Fields {
+			fields += fmt.Sprintf("  %s;\n", field.ProtoType(i+3))
 		}
 
 		// Ensure custom types are imported
-		protoImports := append(opts.Fields.ProtoImports(), opts.Indexes.ProtoImports()...)
-		customFields := append(opts.Fields.Custom(), opts.Indexes.Custom()...)
-		for _, f := range customFields {
+		protoImports := opts.Fields.ProtoImports()
+		for _, f := range opts.Fields.Custom() {
 			protoImports = append(protoImports,
 				fmt.Sprintf("%[1]v/%[2]v/%[3]v.proto", opts.AppName, opts.ModuleName, f),
 			)
@@ -535,26 +397,22 @@ import "%[1]v";`, f)
 
 		templateMessages := `message MsgCreate%[2]v {
   string %[3]v = 1;
-%[4]v
-%[5]v}
+%[4]v}
 message MsgCreate%[2]vResponse {}
 
 message MsgUpdate%[2]v {
   string %[3]v = 1;
-%[4]v
-%[5]v}
+%[4]v}
 message MsgUpdate%[2]vResponse {}
 
 message MsgDelete%[2]v {
   string %[3]v = 1;
-%[4]v}
+}
 message MsgDelete%[2]vResponse {}
-
 %[1]v`
 		replacementMessages := fmt.Sprintf(templateMessages, typed.PlaceholderProtoTxMessage,
 			opts.TypeName.UpperCamel,
 			opts.MsgSigner.LowerCamel,
-			indexes,
 			fields,
 		)
 		content = replacer.Replace(content, typed.PlaceholderProtoTxMessage, replacementMessages)
